Split unit test file
